Data.Functor.WithCounter

Documentation

data WithCounter a

A value with a Natural counter

Constructors

WithCounter

Fields

counter :: !Natural

value :: !a

toWriter :: WithCounter a -> Writer (Sum Natural) a

Convert WithCounter to Writer (Sum Natural).

Used to define the Monad instance

unWriter :: Writer (Sum Natural) a -> WithCounter a

Convert Writer (Sum Natural) to WithCounter.

Used to define the Monad instance

parseNatural :: String -> Parser Natural

Parse a natural number argument, given its field name

parseWithCounter :: Parser a -> Parser (WithCounter a)

Parse a Natural counter and value, given a Parser for the value

toWithCounter :: forall a s. (Natural & (a & s)) :-> (WithCounter a & s)

Wrap WithCounter

unWithCounter :: forall a s. (WithCounter a & s) :-> ((Natural, a) & s)

Unwrap WithCounter

assertWithCounter_ :: (WithCounter a & (Natural & s)) :-> (a & s)

Assert the counter matches the given value

incrementCounter :: (WithCounter a & s) :-> (WithCounter a & s)

Increment the counter